requirements.txt
flask
openpyxl
xlrd
pyinstaller
Pillow
